+++ tests/com/yanncebron/m68kplugin/editor/M68kImplementationTextSelectionerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Authors
+ * Copyright 2026 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,13 +30,96 @@ public class M68kImplementationTextSelectionerTest extends BasePlatformTestCase 
          move.l d0,d1
          endm
          macro<caret>Name""");
-    PsiElement element =
-      TargetElementUtil.findTargetElement(myFixture.getEditor(), TargetElementUtil.getInstance().getAllAccepted());
-    assertNotNull(element);
-    final String implementationText = ImplementationViewComponent.getNewText(element.getNavigationElement());
-    assertEquals("""
+    doTest("""
       macroName macro
        move.l d0,d1
-       endm""", implementationText);
+       endm""");
   }
+
+  public void testLabelFiveFollowingEntries() {
+    myFixture.configureByText("a.s",
+      """
+        label
+          moveq #1,d0
+          moveq #2,d0
+          rept 3
+          endr
+          even
+          moveq #6,d0
+        
+          bra la<caret>bel
+        """);
+    doTest("""
+      label
+        moveq #1,d0
+        moveq #2,d0
+        rept 3
+        endr
+        even""");
+  }
+
+  public void testLocalLabelFiveFollowingEntries() {
+    myFixture.configureByText("a.s",
+      """
+        .localLabel
+          moveq #1,d0
+          moveq #2,d0
+          rept 3
+          endr
+          moveq #5,d0
+          moveq #6,d0
+        
+          bra .localLa<caret>bel
+        """);
+    doTest("""
+      .localLabel
+        moveq #1,d0
+        moveq #2,d0
+        rept 3
+        endr
+        moveq #5,d0""");
+  }
+
+  public void testLabelStopAtFirstLabel() {
+    myFixture.configureByText("a.s",
+      """
+        label
+          moveq #1,d0
+          moveq #2,d0
+        anotherStopLabel
+          moveq #5,d0
+        
+          bra la<caret>bel
+        """);
+    doTest("""
+      label
+        moveq #1,d0
+        moveq #2,d0""");
+  }
+
+  public void testLabelStopAtFirstLocalLabel() {
+    myFixture.configureByText("a.s",
+      """
+        label
+          moveq #1,d0
+          moveq #2,d0
+        .stopLabel
+          moveq #5,d0
+        
+          bra la<caret>bel
+        """);
+    doTest("""
+      label
+        moveq #1,d0
+        moveq #2,d0""");
+  }
+
+  private void doTest(String expectedImplementationText) {
+    PsiElement element = TargetElementUtil.findTargetElement(myFixture.getEditor(),
+      TargetElementUtil.getInstance().getAllAccepted());
+    assertNotNull(element);
+    final String implementationText = ImplementationViewComponent.getNewText(element.getNavigationElement());
+    assertEquals(expectedImplementationText, implementationText);
+  }
+
 }